pyproject.toml
[project]
name = "central-postgres-vps"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
backup = [
    "google-cloud-storage>=2.10.0",
    "apscheduler>=3.10.0",
]
